+++ src/yushkevich/java_2/lesson_1/Jump.java
@@ -1,4 +1,4 @@
-package yushkevich.java_2.lesson_1.HomeWork;
+package yushkevich.java_2.lesson_1;
 
 public interface Jump {
