+++ src/yushkevich/java_2/lesson_1/Main.java
@@ -1,4 +1,4 @@
-package yushkevich.java_2.lesson_1.Work;
+package yushkevich.java_2.lesson_1;
 
 public class Main {
     public static void main(String[] args) {
