+++ src/yushkevich/java_2/lesson_1/Obstacle.java
@@ -1,4 +1,4 @@
-package yushkevich.java_2.lesson_1.Work;
+package yushkevich.java_2.lesson_1;
 
 public abstract class Obstacle {
     protected final int size;
