--- src/yushkevich/java_2/lesson_1/Work/Jump.java
@@ -1,6 +1,0 @@
-package yushkevich.java_2.lesson_1.Work;
-
-public interface Jump {
-
-    void doJump(int height);
-}
