--- src/yushkevich/java_2/lesson_1/Work/Run.java
@@ -1,6 +1,0 @@
-package yushkevich.java_2.lesson_1.Work;
-
-public interface Run {
-
-    void run(int dist);
-}
